+++ logiweb/src/main/java/com/logiweb/avaji/crud/order/service/api/OrderService.java
@@ -44,7 +44,7 @@ public interface OrderService {
      *
      * @param orderId of order entity.
      */
-    void deleteOrder(Integer orderId);
+    void deleteOrder(long orderId);
 
 
 }
