+++ logiweb/src/main/java/com/logiweb/avaji/crud/order/service/implementetion/OrderServiceImpl.java
@@ -66,7 +66,7 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public void deleteOrder(Integer orderId) {
+    public void deleteOrder(long orderId) {
         orderDAO.deleteOrder(orderId);
     }
