+++ logiweb/src/main/java/com/logiweb/avaji/workdetails/service/api/WorkDetailsService.java
@@ -3,8 +3,8 @@ package com.logiweb.avaji.workdetails.service.api;
 import com.logiweb.avaji.workdetails.dto.WorkDetailsDto;
 
 public interface WorkDetailsService {
-    WorkDetailsDto readWorkDetailsByDriverId(Integer driverId);
-    void updateWorkShiftStatus(boolean isActive, Integer workShiftId);
-    void updateDriverStatus(String status, Integer driverId);
-    void updateCargoStatus(String status, Integer cargoId);
+    WorkDetailsDto readWorkDetailsByDriverId(long driverId);
+    void updateWorkShiftStatus(boolean isActive, long workShiftId);
+    void updateDriverStatus(String status, long driverId);
+    void updateCargoStatus(String status, long cargoId);
 }
